skip pg-native on node 26
fixes for pg-native
